:bug: fixed product variant not being handled properly

+++ new/backend/src/main/java/com/kinumna/model/Payment.java
@@ -24,7 +24,7 @@ public class Payment {
     private Integer paymentId;
 
     private LocalDateTime paymentDate;
-    private Float amount;
+    private Integer amount;
 
     @Enumerated(EnumType.STRING)
     private PaymentStatus status;
